Enable byd object no matter the kernel config
It turns out that the BYD code was only being included based on a kernel
config option that the standard Debian kernel was not going to set. I
needed to modify the Makefile to force the inclusion of the module
(after all, if you are building this it's because you want it enabled).

I also had to remove ifdef code that referenced the config option from
byd.h, otherwise I would get complaints about structs being initialized
twice. The code in its current state builds and loads appropriately
against a stock Debian stretch kernel